refactoring .emacs, using macports gpg-agent
in .emacs - clean up the find-file-not-found function;
it is big to keep as a lambda when you need to debug.
Also add asciddoc (not a great mode but...)

In gpg-agent.conf, I've seen flakiness with the
homebrew pinentry client that I didn't see with
the macports version, so cut over.